requirements.txt
discord.py
google-api-python-client
google-auth-oauthlib
apscheduler
python-dotenv
pytz
jishaku
discord.py
asyncpg
termcolor